back end/api-audio/requirements.txt
Flask==3.1.0
whisper==20240930
ffmpeg-python==0.2.0
requests==2.31.0

